pyproject.toml
[project]
name = "ace-skyspark-lib"
version = "0.1.0"
description = "Modern async SkySpark client library with Pydantic validation and Project Haystack compliance"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Andrew Rodgers", email = "andrew@aceiotsolutions.com" }
]
keywords = ["skyspark", "haystack", "iot", "building", "automation", "bms", "async"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Framework :: AsyncIO",
    "Topic :: Home Automation",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.13"
dependencies = [
    "aceiot-models>=0.3.4",
    "pydantic>=2.12.3",
    "httpx>=0.27.0",
    "scramp>=1.4.5",
    "python-dateutil>=2.9.0",
    "pytz>=2024.1",
    "tenacity>=8.5.0",
    "structlog>=24.1.0",
]

[project.scripts]
ace-skyspark-lib = "ace_skyspark_lib:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyrefly>=0.39.4",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.14.2",
    "mypy>=1.13.0",
    "python-dotenv>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "ASYNC", "S", "B", "A", "C4", "DTZ", "RET", "SIM", "ARG", "PTH", "PL", "RUF"]
ignore = ["ANN101", "ANN102", "S101"]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = ["-ra", "--strict-markers", "--strict-config", "-v"]

[tool.pyrefly]
include = ["src/**/*.py"]
